+++ zulu/datetime.py
@@ -4,6 +4,7 @@
 
 from __future__ import absolute_import
 
+import calendar
 from datetime import datetime, timedelta
 from functools import wraps
 import time
@@ -249,6 +250,28 @@ class DateTime(datetime):
             :class:`.DateTime`
         """
         return DateTime(*tuple(self))
+
+    def isleap(self):
+        """Return whether this datetime's :attr:`year` is a leap year.
+
+        Returns:
+            bool
+        """
+        return calendar.isleap(self.year)
+
+    def leapdays(self, year_or_dt):
+        """Return the number of leap days between this datetime's :attr:`year`
+        and `to_year`.
+
+        Args:
+            year_or_dt (int|datetime): Year or year from datetime to calculate
+                leap days between.
+
+        Returns:
+            int
+        """
+        other_year = getattr(year_or_dt, 'year', year_or_dt)
+        return calendar.leapdays(self.year, other_year)
 
     def format(self, format=None, tz=None):
         """Return datetime as a string using the format string `format` while