+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Changelog
 - Add ``DateTime.datetime`` property that returns a native datetime.
 - Add ``DateTime.fromgmtime`` that creates a ``DateTime`` from a UTC based ``time.struct_time``.
 - Add ``DateTime.fromlocaltime`` that creates a ``DateTime`` from a local ``time.struct_time``.
+- Add ``DateTime.isleap`` method that returns whether its year is a leap year.
+- Add ``DateTime.leapdays`` that calculates the number of leap days between its year and another year.
 - Add ``DateTime.add`` and ``DateTime.sub`` methods.
 - Make ``DateTime.shift/add/sub`` use ``dateutil.relativedelta.relativedelta`` under the hood which brings new shiftable arguments: years and months
 - Drop support for milliseconds from ``DateTime.shift/add/sub``. **breaking change**